fix: rename folder themes to theme
