+++ examples/themes/netflix-theme.html
@@ -4,7 +4,7 @@
     <meta name="viewport" content="width=device-width" />
     <title>Media Chrome Netflix Theme Example</title>
     <script type="module" src="../../dist/index.js"></script>
-    <script type="module" src="../../src/js/themes/media-theme.js"></script>
+    <script type="module" src="../../src/js/theme/media-theme.js"></script>
   </head>
   <body>
     <style>
