+++ examples/themes/responsive-theme.html
@@ -7,7 +7,7 @@
     <!-- using a community fork since the original went into maintenance mode -->
     <script src="https://cdn.jsdelivr.net/npm/shadow-container-query-polyfill@1/dist/shadow-container-query-polyfill.modern.js"></script>
     <script type="module" src="../../dist/index.js"></script>
-    <script type="module" src="../../src/js/themes/media-theme.js"></script>
+    <script type="module" src="../../src/js/theme/media-theme.js"></script>
   </head>
   <body>
     <style>
