+++ examples/themes/spotify-theme.html
@@ -3,7 +3,7 @@
   <head>
     <title>Media Chrome Spotify Theme Example</title>
     <script type="module" src="../../dist/index.js"></script>
-    <script type="module" src="../../src/js/themes/media-theme.js"></script>
+    <script type="module" src="../../src/js/theme/media-theme.js"></script>
   </head>
   <body>
     <style>
