+++ examples/themes/youtube-theme.html
@@ -4,7 +4,7 @@
     <meta name="viewport" content="width=device-width" />
     <title>Media Chrome Youtube Theme Example</title>
     <script type="module" src="../../dist/index.js"></script>
-    <script type="module" src="../../src/js/themes/media-theme.js"></script>
+    <script type="module" src="../../src/js/theme/media-theme.js"></script>
     <!-- <script type="module" src="https://cdn.jsdelivr.net/npm/youtube-video-element@0"></script> -->
   </head>
   <body>
